+++ api/sqs-data-consumer/database/models/mark_performance_entries.js
@@ -15,7 +15,7 @@ module.exports = (sequelize, DataTypes) => {
   }
   mark_performance_entries.init({
     page_load_identifier: DataTypes.STRING,
-    duration: DataTypes.INTEGER,
+    duration: DataTypes.DECIMAL,
     entry_type: DataTypes.STRING,
     name: DataTypes.STRING,
     start_time: DataTypes.BIGINT,
